move session manager to runtime common
