+++ runtimes/common/Cargo.toml
@@ -51,6 +51,7 @@ pallet-authorship               = { workspace = true }
 pallet-balances                 = { workspace = true }
 pallet-membership               = { workspace = true }
 pallet-multisig                 = { workspace = true }
+pallet-session                  = { workspace = true }
 pallet-tips                     = { workspace = true }
 pallet-transaction-payment      = { workspace = true }
 pallet-treasury                 = { workspace = true }
@@ -58,6 +59,7 @@ sp-consensus-aura               = { workspace = true }
 sp-core                         = { workspace = true }
 sp-io                           = { workspace = true }
 sp-runtime                      = { workspace = true }
+sp-staking                      = { workspace = true }
 sp-std                          = { workspace = true }
 sp-trie                         = { workspace = true }
 sp-weights                      = { workspace = true }
@@ -128,6 +130,7 @@ std = [
   "pallet-membership/std",
   "pallet-multisig/std",
   "pallet-multisig/std",
+  "pallet-session/std",
   "pallet-tips/std",
   "pallet-transaction-payment/std",
   "pallet-treasury/std",
@@ -164,6 +167,7 @@ try-runtime = [
   "pallet-bonded-coins/try-runtime",
   "pallet-membership/try-runtime",
   "pallet-multisig/try-runtime",
+  "pallet-session/try-runtime",
   "pallet-transaction-payment/try-runtime",
   "pallet-treasury/try-runtime",
   "pallet-treasury/try-runtime",
