+++ runtimes/peregrine/src/system/mod.rs
@@ -27,7 +27,7 @@ use runtime_common::{
 	constants,
 	fees::{ToAuthorCredit, WeightToFee},
 	AccountId, AuthorityId, Balance, BlockHashCount, BlockLength, BlockWeights, FeeSplit, Hash, Nonce,
-	SendDustAndFeesToTreasury, SlowAdjustingFeeUpdate,
+	SendDustAndFeesToTreasury, SessionManager, SlowAdjustingFeeUpdate,
 };
 use sp_core::{ConstBool, ConstU128, ConstU16, ConstU32, ConstU64};
 use sp_runtime::{
@@ -42,9 +42,8 @@ use xcm::v4::Location;
 
 use crate::{
 	governance::{CouncilCollective, RootOrCollectiveProportion, RootOrMoreThanHalfCouncil},
-	weights, Aura, Balances, Block, OriginCaller, PalletInfo, ParachainStaking, PermissionedCollator, Preimage,
-	Runtime, RuntimeCall, RuntimeEvent, RuntimeFreezeReason, RuntimeHoldReason, RuntimeOrigin, RuntimeTask, System,
-	VERSION,
+	weights, Aura, Balances, Block, OriginCaller, PalletInfo, ParachainStaking, Preimage, Runtime, RuntimeCall,
+	RuntimeEvent, RuntimeFreezeReason, RuntimeHoldReason, RuntimeOrigin, RuntimeTask, System, VERSION,
 };
 
 pub(crate) mod proxy;
@@ -142,50 +141,13 @@ impl_opaque_keys! {
 	}
 }
 
-/// The session manager for the collator set.
-pub struct SessionManager;
-
-impl pallet_session::SessionManager<AccountId> for SessionManager {
-	fn new_session(new_index: sp_staking::SessionIndex) -> Option<Vec<AccountId>> {
-		let collators = PermissionedCollator::members().to_vec();
-
-		log::debug!(
-			"assembling new collators for new session {} at #{:?} with {:?}",
-			new_index,
-			System::block_number(),
-			collators
-		);
-
-		System::register_extra_weight_unchecked(
-			<Runtime as frame_system::Config>::DbWeight::get().reads(2),
-			frame_support::pallet_prelude::DispatchClass::Mandatory,
-		);
-
-		if collators.is_empty() {
-			// we never want to pass an empty set of collators. This would brick the chain.
-			log::error!("💥 keeping old session because of empty collator set!");
-			return None;
-		}
-
-		Some(collators)
-	}
-
-	fn start_session(_start_index: sp_staking::SessionIndex) {
-		// We don't care
-	}
-
-	fn end_session(_end_index: sp_staking::SessionIndex) {
-		// We don't care
-	}
-}
-
 impl pallet_session::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type ValidatorId = AccountId;
 	type ValidatorIdOf = ConvertInto;
 	type ShouldEndSession = ParachainStaking;
 	type NextSessionRotation = ParachainStaking;
-	type SessionManager = SessionManager;
+	type SessionManager = SessionManager<Runtime>;
 	type SessionHandler = <SessionKeys as OpaqueKeys>::KeyTypeIdProviders;
 	type Keys = SessionKeys;
 	type WeightInfo = weights::pallet_session::WeightInfo<Runtime>;
